refine dirs and migrate apis
