--- paddle/fluid/pir/serialize_deserialize/include/save_load_parameters.h
@@ -1,42 +1,0 @@
-/* Copyright (c) 2024 PaddlePaddle Authors. All Rights Reserved.
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-    http://www.apache.org/licenses/LICENSE-2.0
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License. */
-#pragma once
-
-#include <string>
-
-#include "paddle/phi/core/dense_tensor.h"
-
-namespace pir {
-
-void SaveFunction(const phi::DenseTensor& x,
-                  const std::string& name,
-                  const std::string& file_path,
-                  bool overwrite,
-                  bool save_as_fp16);
-
-void SaveCombineFunction(const std::vector<const phi::DenseTensor*>& x,
-                         const std::vector<std::string>& names,
-                         const std::string& file_path,
-                         bool overwrite,
-                         bool save_as_fp16,
-                         bool save_to_memory);
-
-void LoadFunction(const std::string& file_path,
-                  int64_t seek,
-                  const std::vector<int64_t>& shape,
-                  bool load_as_fp16,
-                  phi::DenseTensor* out);
-
-void LoadCombineFunction(const std::string& file_path,
-                         const std::vector<std::string>& names,
-                         std::vector<phi::DenseTensor*>* out,
-                         bool load_as_fp16);
-}  // namespace pir
